refactor: extract loadMetaFromCSS
