+++ lib/gulpPlugin.ts
@@ -1,9 +1,6 @@
 import through2 from "through2";
-import fs from "fs";
-import path from "path";
 
-import { generateNoteContent } from "./noteHelper";
-import { StyleMeta } from "../types/styleMeta";
+import { generateNoteContent, loadMetaFromCSS } from "./noteHelper";
 import { Note } from "@hackmd/api/dist/type";
 import API from "@hackmd/api";
 
@@ -19,17 +16,7 @@ export default function ({
   return through2.obj(async function (file, enc, cb) {
     const { contents, path: filePath } = file;
 
-    const meta = JSON.parse(
-      fs.readFileSync(path.resolve(filePath, "../meta.json"), "utf8")
-    ) as StyleMeta;
-
-    if (!meta.slug || !meta.slug.length) {
-      throw new Error("Missing slug in metadata");
-    }
-
-    if (!meta.metadata) {
-      console.warn("Missing style metadata");
-    }
+    const meta = loadMetaFromCSS(filePath);
 
     const note = notesByPermalink[meta.slug];
     const contentToUpdate = generateNoteContent(contents.toString(), meta);
